+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - Added generic PHPDoc annotations to promise APIs
+- Allowed promises to be resolved without passing a value
 
 ### Changed